+++ packages/granularity/src/components/GrList/config.ts
@@ -1,12 +1,9 @@
-import {defineGranularComponent} from '@feugene/unocss-preset-granular/contract'
+import { defineGranularComponent } from '@feugene/unocss-preset-granular/contract'
+
+import { grListSafelist } from './safelist'
 
 export const grListConfig = defineGranularComponent(import.meta.url, {
-    name: 'GrList',
-    safelist: [
-        // GrListItem density
-        'px-4',
-        'py-2',
-        'py-3',
-    ],
-    dependencies: ['GrCard'],
+  name: 'GrList',
+  safelist: grListSafelist,
+  dependencies: ['GrCard', 'GrSkeleton'],
 })